alx_travel_app/alx_travel_app/requirement.txt
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
celery==5.3.4
mysqlclient==2.2.0
python-dotenv==1.0.0
amqp==5.1.1
rabbitmq==0.2.0
pytz==2023.3
Pillow==10.0.0
